+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/service/InstrumentService.java
@@ -2,8 +2,10 @@ package uk.ac.bbsrc.tgac.miso.core.service;
 
 import java.io.IOException;
 import java.util.Collection;
+import java.util.List;
 
 import uk.ac.bbsrc.tgac.miso.core.data.Instrument;
+import uk.ac.bbsrc.tgac.miso.core.data.type.InstrumentType;
 import uk.ac.bbsrc.tgac.miso.core.data.type.PlatformType;
 import uk.ac.bbsrc.tgac.miso.core.util.PaginatedDataSource;
 
@@ -11,6 +13,8 @@ public interface InstrumentService extends DeleterService<Instrument>, ListServi
     SaveService<Instrument> {
 
   Collection<Instrument> listByPlatformType(PlatformType platformType) throws IOException;
+
+  List<Instrument> listByType(InstrumentType type) throws IOException;
 
   Instrument getByName(String name) throws IOException;