+++ miso-web/src/it/java/uk/ac/bbsrc/tgac/miso/webapp/integrationtest/BulkLibraryIT.java
@@ -46,11 +46,12 @@ public class BulkLibraryIT extends AbstractIT {
       LibColumns.CONCENTRATION, LibColumns.CONCENTRATION_UNITS, LibColumns.SPIKE_IN, LibColumns.SPIKE_IN_DILUTION, LibColumns.SPIKE_IN_VOL);
 
   private static final Set<String> editColumns = Sets.newHashSet(LibColumns.SAMPLE_ALIAS, LibColumns.SAMPLE_LOCATION,
-      LibColumns.EFFECTIVE_GROUP_ID, LibColumns.CREATION_DATE, LibColumns.INITIAL_VOLUME, LibColumns.PARENT_NG_USED,
-      LibColumns.PARENT_VOLUME_USED);
+      LibColumns.EFFECTIVE_GROUP_ID, LibColumns.CREATION_DATE, LibColumns.WORKSTATION, LibColumns.THERMAL_CYCLER, LibColumns.INITIAL_VOLUME,
+      LibColumns.PARENT_NG_USED, LibColumns.PARENT_VOLUME_USED);
 
   private static final Set<String> propagateColumns = Sets.newHashSet(LibColumns.SAMPLE_ALIAS, LibColumns.SAMPLE_LOCATION,
-      LibColumns.EFFECTIVE_GROUP_ID, LibColumns.CREATION_DATE, LibColumns.PARENT_NG_USED, LibColumns.PARENT_VOLUME_USED);
+      LibColumns.EFFECTIVE_GROUP_ID, LibColumns.CREATION_DATE, LibColumns.WORKSTATION, LibColumns.THERMAL_CYCLER, LibColumns.PARENT_NG_USED,
+      LibColumns.PARENT_VOLUME_USED);
 
   private static final Set<String> receiptColumns = Sets.newHashSet(SamColumns.SAMPLE_TYPE, SamColumns.SCIENTIFIC_NAME, SamColumns.PROJECT,
       SamColumns.SUBPROJECT, SamColumns.EXTERNAL_NAME, SamColumns.IDENTITY_ALIAS, SamColumns.DONOR_SEX, SamColumns.CONSENT,