+++ RELEASE_NOTES.md
@@ -8,6 +8,9 @@ Changes:
     * Library aliquots: Pool Preparation, Dilution Preparation
   * Sort Transfers list by last modified time
   * Allow only administrators to add tissue materials (detailed sample)
+  * Added workstations
+  * Added thermal cycler instrument type
+  * Allow recording the workstation and thermal cycler used to prepare libraries
   * Fixed an issue where Run Scanner could cause a container to repeatedly flip between multiple
     container models if the container is linked to multiple runs
   * Fixed error when creating Sample Classes